Extract progress listener to custom hook
