+++ src/screens/Game/GamePage/index.tsx
@@ -20,13 +20,7 @@ import UpdateComponent from 'src/components/UI/UpdateComponent'
 
 import { updateGame } from 'src/helpers'
 
-import {
-  AppSettings,
-  GameInfo,
-  GameStatus,
-  InstallInfo,
-  InstallProgress
-} from 'src/types'
+import { AppSettings, GameInfo, GameStatus, InstallInfo } from 'src/types'
 
 import GamePicture from '../GamePicture'
 import TimeContainer from '../TimeContainer'
@@ -39,8 +33,7 @@ import EpicLogo from 'src/assets/epic-logo.svg'
 import GOGLogo from 'src/assets/gog-logo.svg'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faTriangleExclamation } from '@fortawesome/free-solid-svg-icons'
-
-const storage: Storage = window.localStorage
+import { hasProgress } from 'src/hooks/hasProgress'
 
 const { ipcRenderer } = window.require('electron') as {
   ipcRenderer: IpcRenderer
@@ -72,19 +65,8 @@ export default function GamePage(): JSX.Element | null {
   )[0]
 
   const { status } = gameStatus || {}
-  const previousProgress = JSON.parse(
-    storage.getItem(appName) || '{}'
-  ) as InstallProgress
-
+  const [progress, previousProgress] = hasProgress(appName)
   const [gameInfo, setGameInfo] = useState({} as GameInfo)
-  const [progress, setProgress] = useState(
-    previousProgress ??
-      ({
-        bytes: '0.00MiB',
-        eta: '00:00:00',
-        percent: '0.00%'
-      } as InstallProgress)
-  )
   const [autoSyncSaves, setAutoSyncSaves] = useState(false)
   const [savesPath, setSavesPath] = useState('')
   const [isSyncing, setIsSyncing] = useState(false)
@@ -147,22 +129,6 @@ export default function GamePage(): JSX.Element | null {
     }
     updateConfig()
   }, [isInstalling, isPlaying, appName, epicLibrary, gogLibrary])
-
-  useEffect(() => {
-    const onGameStatusUpdate = async (
-      _e: Electron.IpcRendererEvent,
-      { appName: appWithProgress, progress }: GameStatus
-    ) => {
-      if (appName === appWithProgress && progress) {
-        setProgress(progress)
-      }
-    }
-    ipcRenderer.on('setGameStatus', onGameStatusUpdate)
-
-    return () => {
-      ipcRenderer.removeListener('setGameStatus', onGameStatusUpdate)
-    }
-  }, [])
 
   async function handleUpdate() {
     await handleGameStatus({
